requirements.txt
apache-airflow==2.7.1
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
pyarrow==14.0.1
apache-airflow-providers-postgres==5.7.1